first
uploading data (spikeIDtiming.mat, traildata.mat,
20140205…_metadata.mat) + template for iPython notebook + example
figures

From Carl:
	metadata.rt is the sampling rate (Hz). All timings--both for trial
structure and for spike events--are in sample number, so you'll need
this quantity to relate to seconds.
	metadata.numsamples is the total number of samples in the dataset (per
channel, of course). this number divided by rt gives you total time of
the experiment
	trialDataStruct.UOIs lists the ID number of each one of the units
(which is what we call neurons). There are 68 in this dataset.
	trialDataStruct.odorEpochs are the timings (in sample number) of the
onset (column 0) and offset (column 1) of each one of the trials (odor
pulses), of which there were 280 (one trial onset/offset for each row)
	trialDataStruct.valveID is the identity of the odorant on each one of
the 280 trials, in the order in which they were presented (so there is
a straight forward correspondence between valveID and odorEpochs). Note
that 6 is a blank trial (no odorant).
	spikeIDtiming.mat contains the unit ID (column 0) and time of
occurrence (column 1) for all of the spikes in the dataset, of which
there a bit more than half a million.